Implement the pickup function.
1.add some ui, contains the knapsackPanel,SlotPanel,Slot,Item,the grid
layout group component;
2.the script in knapsackPanel,Slot,Item.
The knapsack(Inventory) handle get item:the capacity is 1, the capacity isn't one and have previous
item,the capacity isn't one and...;
The slot handle additem: if handle the first item...else...
The ItemUI handle GUI, update the graphic;
the important is, it cut the function to the right object.
from the requirement to the bottom, maybe can use tdd.